feat: implement ApplicationContext
