+++ nextstep-di/src/test/java/nextstep/di/factory/ConfigurationBeanScannerTest.java
@@ -23,43 +23,6 @@ class ConfigurationBeanScannerTest {
     }
 
     @Test
-    public void register_classpathBeanScanner() {
-        BeanFactory beanFactory = new BeanFactory();
-        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
-        cbs.scanConfigurationBeans();
-
-        ClasspathBeanScanner cbds = new ClasspathBeanScanner(beanFactory);
-        cbds.scanBeans("nextstep.di.factory.example");
-
-        beanFactory.initialize();
-
-        assertNotNull(beanFactory.getBean(DataSource.class));
-
-        JdbcUserRepository userRepository = beanFactory.getBean(JdbcUserRepository.class);
-        assertNotNull(userRepository);
-
-        MyJdbcTemplate jdbcTemplate = beanFactory.getBean(MyJdbcTemplate.class);
-        assertNotNull(jdbcTemplate);
-        assertNotNull(jdbcTemplate.getDataSource());
-    }
-
-    @Test
-    void confirm_di_between_classpath_and_configuration_scanner() {
-        BeanFactory beanFactory = new BeanFactory();
-        ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
-        cbs.scanConfigurationBeans();
-
-        ClasspathBeanScanner cbds = new ClasspathBeanScanner(beanFactory);
-        cbds.scanBeans("nextstep.di.factory.scannerdi");
-
-        beanFactory.initialize();
-
-        assertNotNull(beanFactory.getBean(ExampleRepository.class));
-        assertNotNull(beanFactory.getBean(DataSource.class));
-        assertNotNull(beanFactory.getBean(ExampleRepository.class).getDataSource());
-    }
-
-    @Test
     void appendBasePackages() {
         BeanFactory beanFactory = new BeanFactory();
         ConfigurationBeanScanner cbs = new ConfigurationBeanScanner(beanFactory);
